Group files into packages
