+++ src/amulet/biome/biome.py.cpp
@@ -3,16 +3,18 @@
 #include <pybind11/stl.h>
 #include <pybind11/typing.h>
 
+#include <pybind11_extensions/py_module.hpp>
+
 #include <memory>
 #include <span>
 
-#include <amulet/biome.hpp>
+#include "biome.hpp"
 
 namespace py = pybind11;
 
 void init_biome(py::module m_parent)
 {
-    auto m = m_parent.def_submodule("biome");
+    auto m = pybind11_extensions::def_subpackage(m_parent, "biome");
     py::class_<Amulet::Biome, Amulet::PlatformVersionContainer, std::shared_ptr<Amulet::Biome>> Biome(m, "Biome",
         "A class to manage the state of a biome.\n"
         "\n"
