+++ src/amulet/block/block.hpp
@@ -7,7 +7,7 @@
 #include <amulet/dll.hpp>
 #include <amulet/io/binary_reader.hpp>
 #include <amulet/io/binary_writer.hpp>
-#include <amulet/version.hpp>
+#include <amulet/version/version.hpp>
 #include <amulet_nbt/tag/int.hpp>
 #include <amulet_nbt/tag/string.hpp>
