+++ src/amulet/block_entity/block_entity.hpp
@@ -7,19 +7,16 @@
 #include <amulet/dll.hpp>
 #include <amulet/io/binary_reader.hpp>
 #include <amulet/io/binary_writer.hpp>
-#include <amulet/version.hpp>
+#include <amulet/version/version.hpp>
 #include <amulet_nbt/tag/eq.hpp>
 #include <amulet_nbt/tag/named_tag.hpp>
 
 namespace Amulet {
-class Entity : public PlatformVersionContainer {
+class BlockEntity : public PlatformVersionContainer {
 private:
     std::string _namespace;
     std::string _base_name;
     std::shared_ptr<AmuletNBT::NamedTag> _nbt;
-    double _x;
-    double _y;
-    double _z;
 
 public:
     AMULET_CORE_EXPORT const std::string& get_namespace() const;
@@ -31,29 +28,16 @@ public:
     AMULET_CORE_EXPORT std::shared_ptr<AmuletNBT::NamedTag> get_nbt() const;
     AMULET_CORE_EXPORT void set_nbt(std::shared_ptr<AmuletNBT::NamedTag> nbt);
 
-    AMULET_CORE_EXPORT double get_x() const;
-    AMULET_CORE_EXPORT double get_y() const;
-    AMULET_CORE_EXPORT double get_z() const;
-
-    AMULET_CORE_EXPORT void set_x(double);
-    AMULET_CORE_EXPORT void set_y(double);
-    AMULET_CORE_EXPORT void set_z(double);
-
-    AMULET_CORE_EXPORT Entity(
-        PlatformType platform,
-        VersionNumber version,
-        std::string namespace_,
-        std::string base_name,
-        double x,
-        double y,
-        double z,
+    AMULET_CORE_EXPORT BlockEntity(
+        const PlatformType& platform,
+        const VersionNumber& version,
+        const std::string& namespace_,
+        const std::string& base_name,
         std::shared_ptr<AmuletNBT::NamedTag> nbt);
 
-    AMULET_CORE_EXPORT ~Entity();
-
     AMULET_CORE_EXPORT void serialise(BinaryWriter&) const;
-    AMULET_CORE_EXPORT static Entity deserialise(BinaryReader&);
+    AMULET_CORE_EXPORT static BlockEntity deserialise(BinaryReader&);
 
-    AMULET_CORE_EXPORT bool operator==(const Entity& other) const;
+    AMULET_CORE_EXPORT bool operator==(const BlockEntity& other) const;
 };
 }
