+++ src/amulet/entity/__init__.pyi
