+++ src/amulet/level/java/chunk_encode.cpp
@@ -5,7 +5,7 @@
 
 #include <amulet_nbt/tag/named_tag.hpp>
 
-#include <amulet/chunk.hpp>
+#include <amulet/chunk/chunk.hpp>
 
 #include "chunk.hpp"
 #include "raw_dimension.hpp"
