+++ src/amulet/version/__init__.pyi
