+++ src/amulet/version/version.hpp
